+++ src/notyf/index.js
@@ -1,6 +1,6 @@
 import { has_dom } from '@ctx-core/dom'
-import { notyf_config } from '../notyf_config/index.js'
 import { Notyf, NotyfNotification } from 'notyf'
+import { notyf_config } from '../notyf_config/index.js'
 export const notyf =
 	has_dom
 	? new Notyf(notyf_config)