+++ src/notyf_error/index.d.ts
@@ -1,6 +1,6 @@
-import { be_atom_triple_T } from '@ctx-core/nanostores'
 import type { WritableAtom_ } from '@ctx-core/nanostores'
-import type { Be, Ctx } from '@ctx-core/object'
+import { be_atom_triple_T } from '@ctx-core/nanostores'
+import type { Ctx } from '@ctx-core/object'
 import type { INotyfNotificationOptions, NotyfNotification } from 'notyf'
 export declare function notyf_error(
 	ctx:Ctx, in_payload_arg:string|Partial<INotyfNotificationOptions>