+++ src/notyf_options_/index.d.ts
@@ -1,2 +1,4 @@
 import type { INotyfNotificationOptions } from 'notyf'
-export declare function notyf_options_(in_options:string|Partial<INotyfNotificationOptions>):INotyfNotificationOptions
+export declare function notyf_options_(
+	in_options:string|Partial<INotyfNotificationOptions>
+):INotyfNotificationOptions