+++ src/notyf_warn/index.d.ts
@@ -1,7 +1,14 @@
-import type { WritableAtom_ } from '@ctx-core/nanostores'
-import type { Be, Ctx } from '@ctx-core/object'
+import type { be_atom_triple_T } from '@ctx-core/nanostores'
+import type { Ctx } from '@ctx-core/object'
 import type { INotyfNotificationOptions, NotyfNotification } from 'notyf'
+export declare const [
+	notyf_warn__backgroundColor$_,
+	notyf_warn__backgroundColor_,
+	notyf_warn__backgroundColor__set,
+]:be_atom_triple_T<string>
+export {
+	notyf_warn__backgroundColor$_ as notyf_warn__backgroundColor__,
+}
 export declare function notyf_warn(
 	ctx:Ctx, in_payload_arg:string|Partial<INotyfNotificationOptions>
 ):Promise<NotyfNotification>
-export const notyf_warn__backgroundColor__:Be<WritableAtom_<string>>